Refactor model components and enhance dataset validation
- Updated FrozenBatchNorm2d to handle num_batches_tracked during state loading.
- Improved documentation in decoder and encoder modules for clarity.
- Enhanced HungarianMatcher to ensure float precision during probability calculations.
- Removed outdated Colab notebooks for training and inference; replaced with updated versions.
- Added a new script to check dataset integrity before training.
- Improved instructions for organizing the OPIXray dataset.
- Simplified overfitting script by removing synthetic mode, focusing on real data.

+++ models/matcher.py
@@ -21,11 +21,14 @@ class HungarianMatcher(nn.Module):
     @torch.no_grad()
     def forward(self, outputs: Dict, targets: List[Dict]) -> List[Tuple[torch.Tensor, torch.Tensor]]:
         B, Q = outputs["pred_logits"].shape[:2]
-        out_prob = outputs["pred_logits"].flatten(0, 1).sigmoid()   # [B*Q, C]
-        out_bbox = outputs["pred_boxes"].flatten(0, 1)              # [B*Q, 4]
+        # fp32 throughout: under AMP the logits arrive as fp16, where sigmoid() of a
+        # confidently-negative logit underflows to 0 and the 1e-8 floor below is itself
+        # subnormal -> log(0) = -inf poisons the cost matrix and breaks the assignment.
+        out_prob = outputs["pred_logits"].float().flatten(0, 1).sigmoid()   # [B*Q, C]
+        out_bbox = outputs["pred_boxes"].float().flatten(0, 1)              # [B*Q, 4]
 
         tgt_ids = torch.cat([t["labels"] for t in targets])
-        tgt_bbox = torch.cat([t["boxes"] for t in targets])
+        tgt_bbox = torch.cat([t["boxes"] for t in targets]).float()
 
         if tgt_ids.numel() == 0:
             return [(torch.as_tensor([], dtype=torch.int64),
